Moved filelist handling to separate helper functions.
git-svn-id: svn://rootdirectory.ddns.net/pdclib/trunk@808 bcf39385-58cc-4174-9fcf-14f50f90dd47
