+++ functions/stdio/fopen.c
@@ -84,10 +84,7 @@ struct _PDCLIB_file_t * fopen( const char * _PDCLIB_restrict filename, const cha
     rc->status |= _IOLBF;
     /* TODO: Setting mbstate */
     /* Adding to list of open files */
-    _PDCLIB_LOCK( _PDCLIB_filelist_mtx );
-    rc->next = _PDCLIB_filelist;
-    _PDCLIB_filelist = rc;
-    _PDCLIB_UNLOCK( _PDCLIB_filelist_mtx );
+    _PDCLIB_setstream( rc );
     return rc;
 }
