+++ platform/example/functions/stdio/tmpfile.c
@@ -102,10 +102,7 @@ struct _PDCLIB_file_t * tmpfile( void )
     rc->ungetidx = 0;
     /* TODO: Setting mbstate */
     /* Adding to list of open files */
-    _PDCLIB_LOCK( _PDCLIB_filelist_mtx );
-    rc->next = _PDCLIB_filelist;
-    _PDCLIB_filelist = rc;
-    _PDCLIB_UNLOCK( _PDCLIB_filelist_mtx );
+    _PDCLIB_setstream( rc );
     return rc;
 }
